Move 11ty utils to their own folder
